[bug-fix] info time is double
git-svn-id: http://svn.code.sf.net/p/rts-2/code/branches/rts-2/REL_0_5_0_exp@3764 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/utils/rts2daemon.cpp
@@ -723,7 +723,8 @@ Rts2Daemon::info ()
 {
   struct timeval infot;
   gettimeofday (&infot, NULL);
-  info_time->setValueDouble (infot.tv_sec + infot.tv_usec / USEC_SEC);
+  info_time->setValueDouble (infot.tv_sec +
+			     (double) infot.tv_usec / USEC_SEC);
   return 0;
 }
